many tests

+++ XL.API/Features/Expressions/GetSheetCellValueQueryHandler.cs
@@ -3,8 +3,8 @@ using XL.API.Models;
 
 namespace XL.API.Features.Expressions;
 
-public record GetSheetCellValueQuery(string SheetId, string CellId) : IRequest<OneOf<string, double, NotFound>>;
-public class GetSheetCellValueQueryHandler : IRequestHandler<GetSheetCellValueQuery, OneOf<string, double, NotFound>>
+public record GetSheetCellValueQuery(string SheetId, string CellId) : IRequest<OneOf<string, decimal, NotFound>>;
+public class GetSheetCellValueQueryHandler : IRequestHandler<GetSheetCellValueQuery, OneOf<string, decimal, NotFound>>
 {
     private readonly ISheetCellRepository repository;
 
@@ -13,7 +13,7 @@ public class GetSheetCellValueQueryHandler : IRequestHandler<GetSheetCellValueQu
         this.repository = repository;
     }
     
-    public async Task<OneOf<string, double, NotFound>> Handle(GetSheetCellValueQuery request, CancellationToken cancellationToken)
+    public async Task<OneOf<string, decimal, NotFound>> Handle(GetSheetCellValueQuery request, CancellationToken cancellationToken)
     {
         var cell = await repository.Find(request.SheetId, request.CellId);
         if (cell == null)
